web/storagenode: gross total added to historical data, with surge moved
WHAT:
changed estimation table row order.

WHY:
to show gross total for selected period to avoid misunderstanding
when held amount is bigger than paid multiple times.

+++ storagenode/payout/service.go
@@ -347,7 +347,7 @@ func parsePeriodRange(periodStart, periodEnd string) (periods []string, err erro
 }
 
 // GetHeldRate returns held rate for specific period from join date of node.
-func GetHeldRate(joinTime time.Time, requestTime time.Time) (heldRate int64) {
+func GetHeldRate(joinTime time.Time, requestTime time.Time) (heldRate float64) {
 	monthsSinceJoin := date.MonthsBetweenDates(joinTime, requestTime)
 	switch monthsSinceJoin {
 	case 0, 1, 2:
